Rename Header components to PanelHeader
