+++ src/components/constraints/ConstraintsPanel.svelte
@@ -8,9 +8,9 @@
   import effects from '../../utilities/effects';
   import GridMenu from '../menus/GridMenu.svelte';
   import CssGrid from '../ui/CssGrid.svelte';
-  import HeaderActionButton from '../ui/HeaderActionButton.svelte';
-  import HeaderActions from '../ui/HeaderActions.svelte';
   import Panel from '../ui/Panel.svelte';
+  import PanelHeaderActionButton from '../ui/PanelHeaderActionButton.svelte';
+  import PanelHeaderActions from '../ui/PanelHeaderActions.svelte';
   import ConstraintListItem from './ConstraintListItem.svelte';
 
   export let gridId: number;
@@ -32,9 +32,9 @@
 <Panel>
   <svelte:fragment slot="header">
     <GridMenu {gridId} title="Constraints" />
-    <HeaderActions status={$checkConstraintsStatus}>
-      <HeaderActionButton title="Check Constraints" showLabel on:click={() => effects.checkConstraints()} />
-    </HeaderActions>
+    <PanelHeaderActions status={$checkConstraintsStatus}>
+      <PanelHeaderActionButton title="Check Constraints" showLabel on:click={() => effects.checkConstraints()} />
+    </PanelHeaderActions>
   </svelte:fragment>
 
   <svelte:fragment slot="body">
