+++ src/components/simulation/SimulationPanel.svelte
@@ -14,9 +14,9 @@
   import { getArguments, getFormParameters } from '../../utilities/parameters';
   import GridMenu from '../menus/GridMenu.svelte';
   import Parameters from '../parameters/Parameters.svelte';
-  import HeaderActionButton from '../ui/HeaderActionButton.svelte';
-  import HeaderActions from '../ui/HeaderActions.svelte';
   import Panel from '../ui/Panel.svelte';
+  import PanelHeaderActionButton from '../ui/PanelHeaderActionButton.svelte';
+  import PanelHeaderActions from '../ui/PanelHeaderActions.svelte';
 
   export let gridId: number;
 
@@ -63,9 +63,9 @@
 <Panel>
   <svelte:fragment slot="header">
     <GridMenu {gridId} title="Simulation" />
-    <HeaderActions status={$simulationStatus}>
-      <HeaderActionButton title="Simulate" showLabel on:click={() => effects.simulate()} />
-    </HeaderActions>
+    <PanelHeaderActions status={$simulationStatus}>
+      <PanelHeaderActionButton title="Simulate" showLabel on:click={() => effects.simulate()} />
+    </PanelHeaderActions>
   </svelte:fragment>
 
   <svelte:fragment slot="body">
