+++ src/components/ui/PanelHeaderActionButton.svelte
