+++ src/components/ui/PanelHeaderActions.svelte
